requirements.txt
blinker==1.9.0
camelot-py==1.0.9
cffi==2.0.0
chardet==5.2.0
charset-normalizer==3.4.3
click==8.3.0
cryptography==46.0.2
et_xmlfile==2.0.0
Flask==3.1.2
gunicorn==23.0.0
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.3
numpy==2.2.6
opencv-python-headless==4.12.0.88
openpyxl==3.1.5
packaging==25.0
pandas==2.3.3
pdfminer.six==20250506
pdfplumber==0.11.7
pillow==11.3.0
pycparser==2.23
pypdf==3.17.4
pypdfium2==4.30.0
python-dateutil==2.9.0.post0
pytz==2025.2
six==1.17.0
tabulate==0.9.0
tzdata==2025.2
Werkzeug==3.1.3
